Updated 9/4/2017

+++ CodeChef/CodeChef_FLOW011.cpp
@@ -5,12 +5,12 @@ int main() {
 	int t;
 	scanf("%d",&t);
 	while(t --) {
-		int n;
-		scanf("%d",&n);
+		double n;
+		scanf("%lf",&n);
 		if(n < 1500)
-			printf("%d\n",int(n+0.1*n+0.9*n));
+			printf("%.6g\n",n+0.1*n+0.9*n);
 		else
-			printf("%.1f\n",n+500+0.98*n);
+			printf("%.6g\n",n+500+0.98*n);
 	}
 
 	return 0;
