+++ CodeChef/CodeChef_OJUMPS.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-	int n;
-	scanf("%d",&n);
+	long long int n;
+	scanf("%lld",&n);
 
 	if(n%6 == 0 || (n-3)%6 == 0 || (n-1)%6 == 0)
 		printf("yes\n");
